搜索时间转双浮点 增加搜索速度的属性

+++ src/main/java/org/yanbwe/searchcarefully/client/ClientOverlayRenderer.java
@@ -88,7 +88,7 @@ public class ClientOverlayRenderer {
                     if (slot.hasItem()) {
                         ItemStack stack = slot.getItem();
                         if (ItemStackHelper.hasRemainingSearchTime(stack)) {
-                            int searchTime = ItemStackHelper.getRemainingSearchTime(stack);
+                            int searchTime = (int) ItemStackHelper.getRemainingSearchTime(stack);
                             if (searchTime > 0) {
                                 activeSearchSlots.add(slot);
                             }
